Shiny Application of a conversation landscape.

Multiple tabs - one for cleaning and selecting, others for summary plots.

conversation_landscape(
  data,
  ...,
  id = document,
  text_var = text,
  colour_var,
  cleaned_text_var = clean_text,
  date_var = date,
  sentiment_var = sentiment,
  url_var = permalink,
  size = 2,
  x_var = V1,
  y_var = V2,
  type = "scattergl",
  colour_mapping = NULL
)

Arguments

data

Data Frame

...

Additional Columns to select

id

Your ID column, if there is not one in your data, create one

text_var

The original text variable for display in the data table

colour_var

The variable you wish to map colour to (should be a string or factor)

cleaned_text_var

The cleaned text variable for bigram and token plots

date_var

Your date variable

sentiment_var

Sentiment variable

url_var

Variable which contains a url (preferably after using link_click)

size

x_var

Variable which contains your x co-ordinates

y_var

Variable which contains your y co-ordinates

type

Type of the plotly output, unlikely to change from 'scattergl'

colour_mapping

Which colours the colour variable should be mapped to

Value

A shiny application